Define task id type in prefix of id string.
Summary:
This introduces the `TaskIdWithIncarnation` id type. It is similar to
`ResidentTaskId` except that it prefixes the Mesos task id string with
`$tiwi$`. This makes parsing ids easier in the future, eg if we add the
version to the task id string with could introduce `TaskIdWithVersion`
with the type prefix `$tiv$`.

JIRA issues: MARATHON-8501